Aroon benchmark fixes

+++ runtime/src/chpl-cache-support.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2014 Cray Inc.
+ * Copyright 2004-2015 Cray Inc.
  * Other additional copyright holders may be indicated within.
  * 
  * The entirety of this work is licensed under the Apache License,
@@ -293,7 +293,7 @@ void set_valids_for_skip_len(uint64_t* valid, uintptr_t skip, uintptr_t len, int
       // entirely after the region -- nothing valid
     } else if( skip <= offset && offset + 64 <= skip+len ) {
       // entirely within the region -- totally valid
-      valid[i] = (uintptr_t) -1;
+      valid[i] = (uint64_t) -1;
     } else {
       use_start = skip - offset;
       if( use_start < 0 ) use_start = 0;
